migrated graphql functionality to main_graphqlController.js file
